+++ packages/instrumentation-console/src/index.ts
@@ -1,0 +1,7 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+export { ConsoleInstrumentation } from './instrumentation';
+export type { ConsoleInstrumentationConfig } from './types';